ASW-288 Refactor used main shop and fallback config rules
