+++ Resources/services/rules.xml
@@ -5,6 +5,9 @@
     <services>
 
         <service id="AdyenPayment\Rule\AdyenApi\UsedMainShopConfigRule">
+            <argument type="service" id="AdyenPayment\Rule\AdyenApi\UsedFallbackConfigRule"/>
+        </service>
+        <service id="AdyenPayment\Rule\AdyenApi\UsedFallbackConfigRule">
             <argument type="expression">service('models').getRepository('Shopware\\Models\\Shop\\Shop')</argument>
             <argument type="service" id="AdyenPayment\Rule\AdyenApi\IsMainShopConfigRule"/>
         </service>
